moved resources under installs
