+++ services/platform/operator/controller/installs/resources/operator.go
