+++ services/platform/operator/controller/installs/resources/resources.go
