+++ services/platform/operator/controller/installs/resources/talos.go
